+++ RELEASES.md
@@ -1,10 +1,12 @@
-Version tw-0.8.9-3.5 (2021-11-08)
+Version tw-0.8.9-3.5 (2021-11-10)
 ========================
  * NEW: Latest 0.8.10 supercombo (laneless + laneful) model
  * NEW: Latest 0.8.10 driver monitoring model
  * NEW: Update to AGNOS2 for Comma3 users
  * NEW: Added follow level UI metric
- * IMPROVED: Add "+" after max speed to indicate that coasting is enabled (ring around brake indicator also shown)
+ * IMPROVED: Toggle coast mode while driging by tapping max speed (indicated by "+" after max speed; ring around brake indicator no longer shown)
+ * IMPROVED: Toggle one-pedal engage on gas feature while driving by tapping pedal icon (indicated by ring around pedal icon)
+ * IMPROVED: Toggle auto-pause of auto-steer (when in one-pedal mode below 30mph with blinker) while driving by tapping wheel icon (indicated by ring around wheel icon while one-pedal mode active)
  * IMPROVED: Better (but still hackey) follow braking when coasting behind lead downhill
  * FIXED: After using one-pedal hard braking during a driving session, it no longer defaults to moderate braking for future one-pedal braking engagements
  * FIXED: JShuler panda gm steering fix